Remove alignment flags when also using the wx.EXPAND flag (#678)
If a window in a sizer is expanded, then all the alignment flags are
ignored (the alignment flags affect the window position, not the
contents of the window).  wxWidgets simply ignores those flags but
since wxPython 4.1 these raise AssertionError.

This commit removes some of the alignment flags that are being used in
conjunction with the wx.EXPAND flag.